refactor(registry): recast domain-specific cloud widgets as Cloud blocks
Adopt a clearer components-vs-blocks boundary: the components tier holds
controls reusable across unrelated domains, while a widget that only makes
sense for one domain becomes a block under a kind that names that domain.

Apply it to the cloud set (the pilot for this split):

- New "cloud" block kind. Ten infra-specific items move from components to
  blocks: server-card, vm-table, k8s-pod-table, resource-status,
  cluster-map, network-topology, storage-browser, log-viewer, terminal,
  deployment-history. Each block file keeps its data-slot parts as named
  exports and adds a default composition used as the block preview.
- Two genuinely generic controls stay components: metric-card (data
  display) and yaml-editor (inputs).

Move the source into blocks/<name>/, drop the per-component example demos
(the block file previews itself through /embed), swap the example loaders
for block loaders, and regenerate registry.json, props and redirects.
Document the boundary in docs/README.md and docs/catalog.md.
