+++ registry/hirael/blocks/log-viewer/log-viewer.tsx
@@ -109,3 +109,47 @@ function LogLine({
 }
 
 export { LogViewer, LogLine };
+
+export default function LogViewerBlock() {
+  return (
+    <section
+      data-slot="log-viewer-block"
+      className="flex w-full justify-center bg-background p-6 sm:p-10"
+    >
+      <div className="w-full max-w-2xl" dir="ltr">
+        <LogViewer>
+          <LogLine time="12:04:01" level="info" source="[api]">
+            listening on :8080
+          </LogLine>
+          <LogLine time="12:04:01" level="debug" source="[db]">
+            pool acquired (8 idle, 2 active)
+          </LogLine>
+          <LogLine time="12:04:02" level="success" source="[api]">
+            GET /health 200 3ms
+          </LogLine>
+          <LogLine time="12:04:03" level="info" source="[worker]">
+            job 4821 started (resize-image)
+          </LogLine>
+          <LogLine time="12:04:03" level="warn" source="[cache]">
+            key eviction: memory at 82%
+          </LogLine>
+          <LogLine time="12:04:04" level="info" source="[api]">
+            POST /v1/deploy 202 41ms
+          </LogLine>
+          <LogLine time="12:04:05" level="error" source="[worker]">
+            job 4821 failed: upstream timeout after 30s
+          </LogLine>
+          <LogLine time="12:04:05" level="warn" source="[worker]">
+            retry 1/3 scheduled in 2s
+          </LogLine>
+          <LogLine time="12:04:07" level="success" source="[worker]">
+            job 4821 recovered on retry
+          </LogLine>
+          <LogLine time="12:04:08" level="debug" source="[db]">
+            slow query 214ms: SELECT * FROM events
+          </LogLine>
+        </LogViewer>
+      </div>
+    </section>
+  );
+}
