+++ registry/hirael/blocks/vm-table/vm-table.tsx
@@ -194,3 +194,105 @@ export {
   VmTableName,
   VmStatus,
 };
+
+const VM_ROWS: {
+  name: string;
+  id: string;
+  state: VmState;
+  size: string;
+  region: string;
+  ip: string;
+  uptime: string;
+}[] = [
+  {
+    name: "web-prod-01",
+    id: "i-0a1b2c3d",
+    state: "running",
+    size: "c6i.2xlarge",
+    region: "us-east-1",
+    ip: "10.0.1.24",
+    uptime: "42d",
+  },
+  {
+    name: "worker-prod-02",
+    id: "i-0e4f5a6b",
+    state: "running",
+    size: "c6i.xlarge",
+    region: "us-east-1",
+    ip: "10.0.1.51",
+    uptime: "42d",
+  },
+  {
+    name: "batch-runner-07",
+    id: "i-0c7d8e9f",
+    state: "starting",
+    size: "m6i.large",
+    region: "eu-west-2",
+    ip: "10.1.2.13",
+    uptime: "—",
+  },
+  {
+    name: "db-replica-03",
+    id: "i-0b3c4d5e",
+    state: "error",
+    size: "r6i.4xlarge",
+    region: "eu-west-2",
+    ip: "10.1.2.44",
+    uptime: "—",
+  },
+  {
+    name: "sandbox-14",
+    id: "i-0f6a7b8c",
+    state: "stopped",
+    size: "t3.medium",
+    region: "ap-south-1",
+    ip: "—",
+    uptime: "—",
+  },
+];
+
+export default function VmTableBlock() {
+  return (
+    <section
+      data-slot="vm-table-block"
+      className="flex w-full justify-center bg-background p-6 sm:p-10"
+    >
+      <div className="w-full max-w-3xl">
+        <VmTable caption="5 instances across 3 regions">
+          <VmTableHeader>
+            <VmTableRow>
+              <VmTableHead>Instance</VmTableHead>
+              <VmTableHead>Status</VmTableHead>
+              <VmTableHead>Size</VmTableHead>
+              <VmTableHead>Region</VmTableHead>
+              <VmTableHead>IP</VmTableHead>
+              <VmTableHead className="text-end">Uptime</VmTableHead>
+            </VmTableRow>
+          </VmTableHeader>
+          <VmTableBody>
+            {VM_ROWS.map((row) => (
+              <VmTableRow key={row.id}>
+                <VmTableName id={row.id}>{row.name}</VmTableName>
+                <VmTableCell>
+                  <VmStatus state={row.state} />
+                </VmTableCell>
+                <VmTableCell className="font-mono text-xs text-muted-foreground">
+                  {row.size}
+                </VmTableCell>
+                <VmTableCell className="text-muted-foreground">
+                  {row.region}
+                </VmTableCell>
+                <VmTableCell className="font-mono text-xs text-muted-foreground">
+                  {row.ip}
+                </VmTableCell>
+                <VmTableCell className="text-end font-mono text-xs text-muted-foreground">
+                  {row.uptime}
+                </VmTableCell>
+              </VmTableRow>
+            ))}
+          </VmTableBody>
+        </VmTable>
+      </div>
+    </section>
+  );
+}
